Modal API Changes

+++ src/actions/facultyAction.js
@@ -22,8 +22,8 @@ export const addFaculty = (faculty, history) => async (dispatch) => {
 /* Get Faculty */
 export const getFaculty = (id, history) => async (dispatch) => {
   try {
-    const res = axios.get(`/faculty/${id}`);
-    console.log(res);
+    const res = await axios.get(`/faculty/${id}`);
+
     dispatch({
       type: GET_FACULTY,
       payload: res.data,
@@ -52,13 +52,38 @@ export const getFaculties = () => async (dispatch) => {
   }
 };
 
+/* Update Faculty */
+export const updateFaculty = (id, updatedFaculty, history) => async (
+  dispatch
+) => {
+  try {
+    await axios.put(`/faculty/${id}`, updatedFaculty);
+    history.push("/facultyDashboard");
+    dispatch({
+      type: GET_ERROR,
+      payload: {},
+    });
+  } catch (err) {
+    dispatch({
+      type: GET_ERROR,
+      payload: err.response.data,
+    });
+  }
+};
+
 /* Delete a Faculty */
 export const deleteFaculty = (id) => async (dispatch) => {
   try {
-    await axios.delete(`/faculty/${id}`);
-    dispatch({
-      type: DELETE_FACULTY,
-      payload: id,
-    });
+    if (
+      window.confirm(
+        "Are you sure? This will delete all the data related to it"
+      )
+    ) {
+      await axios.delete(`/faculty/${id}`);
+      dispatch({
+        type: DELETE_FACULTY,
+        payload: id,
+      });
+    }
   } catch (err) {}
 };
